Guard against re-installing Kafka
It would be potentially devastating if a new major version of Kafka was 
installed, so this guards against that. One could also argue that the
guards shouldn't be present and that it is up to the user to make sure
that new versions aren't installed uncontrolled. 
To me this feels like a good enough solution for now.