Update attributes related to configuration of Kafka broker
There were a whole slew of new attributes introduced in the final release of v0.8.0, as well as a few removed (metrics related).
Also handles 'map' based attributes, i.e. per topic specific attributes.

Not too happy about some of the attribute naming, so might change that in a near future, but other than that this should be good to go.