Add hackyish support for building Kafka v0.8.1 from source
Maaaan, this fucking mess I tell you. v0.8.1 switched from 'sbt' to 
'gradle', and it was just a cluster fuck to actually construct a
'release' tar.gz archive. Google is my best friend, always and forever.

I'm not happy about this current solution, and I'm not done with it yet.
There's plenty of room for improvement and making things pretty.
Just you wait.