Change all `configuration` attributes to `nil`
Kafka does a pretty good job at keeping track of default values on it's
own, and I see no point in attempting to keep default values consistent
among different versions of Kafka. That's up to them to manage.

Everything is still highly configurable, but with the sane defaults that
the Kafka team decides on.