Test different versions of Chef in Travis
The README points out that Chef >= 11.8.2 is supported, though I have
realised that it actually works with 11.6.0 (and probably any other
patch versions, if there are any). It does however not work with Chef
11.4.4 as it is using some features of Chef that was introduced in
11.6.0.

There was recently an issue (#37) opened regarding this cookbook not
working on OpsWorks (which uses 11.4.4 of Chef), and I might consider
adding support for 11.4.4, though it depends on how much work is needed
to get it in place (not too much probably). For now I’ll include 11.4.4
in `.travis.yml` in case I (or someone else) decide to start adding
support.