Add support for systemd as an `init_style`
First stab at getting `systemd` to work. Still not sure about a few
things such as using an EnvironmentFile and the code for setting up
everything related to init styles is even messier, so it could probably
use some cleanup and rethinking.

Also added a`systemd?` helper for integration tests, as the current
version of serverspec that is used is unable to check if a service is
enbaled for distros that use systemd, so until there's a better fix
this'll have to to.