Generate property keys from `broker` attributes
Instead of defining _ALL_ possible configuration attributes in
`attributes/default.rb` we’ll generate key-value pair from whatever
attributes that are found under the `broker` namespace.

The diffs in this commit are so _HORRIBLE_, but I couldn’t think of any
other way to get this done. There are still some tests that needs to be
rewritten / added, and I do believe that the configuration generation
creates some unnecessary whitespace / newlines that needs to be dealt
with, but everything has it’s time, mkay?

This is going to be so much easier to maintain.