Version bump to v0.4.1
I’d consider this a patch release due to the fact that this should’ve
been done since the start, and I’m a bit surprised that no-one has filed
an issue about it. Suppose that’s because Kafka actually create the log
directories and then it’s not a big deal.