Allow `dotted` notation for broker configuration
This commit makes it possible to use either underscored symbol keys for
configuration attributes (i.e. attributes defined under `broker`), or to
use a `dotted` String notation, i.e.

```
node.set.broker = {‘log.dirs’ => %w[/some/dir-1 /some/dir-2]}
```

Decided to get rid of support for using nested Hashes as it’ll become
difficult to render Hash-based options (CSV strings) correctly, and due
to the naming of some of the configuration parameters it’s not possible
to use `node.default.this.attribute = ‘something’` as the parameters
include `default`, `fetch` and whatever else.

Removed a lot of unit tests that are unnecessary now that the broker
configuration file is dynamically rendered.