Configurable log4j configuration
When I deployed a Kafka cluster the other week I realised that some of
the “default” options in the bundled log4j configuration file included
TRACE level, which is way too verbose to use in production (as it most
likely will affect performance).

Rather than just updating the log4j template I figured I could as well
make the log4j configuration a bit more configurable, so to speak. The
configuration is a bit verbose, but it’s what I could come up with right
now. Will most likely revisit this in the future.

Should be noted that I’ve kept the TRACE level for a few loggers, but
that’s mostly for historic purposes and will be updated in a later
commit.