Remove setup of ZooKeeper
Setting up ZooKeeper is well beyond the scope of this cookbook, as it’s
only intention is to set up Kafka brokers. It’s also _NOT_ recommended
to set up ZooKeeper on the same hardware that Kafka is running on, for
quite obvious reasons.

Maintaining the current / previous ZooKeeper while this cookbook is
transitioning to a more dynamic approach to configuration files is going
to be a lot more difficult and most likely introduce complexity and
exceptions that wouldn’t be needed if this cookbook only set up Kafka.