Switch TRACE level with INFO
I wouldn’t deploy anything with TRACE / DEBUG as log level to production
as it will *most likely* affect performance. I know that the log4j file
that is bundled with Kafka defines these at TRACE, but I don’t really
think that the file is intended to be used in a production environment
to be honest.