Respect `controlled.shutdown.enable` in init scripts
Basically don’t force-kill the Kafka process if `controlled.shutdown.
enable` is set to true. While it’s not necessarily a problem to force-
kill the process, it does take quite some time to verify all the logs
when it starts up again, and sort of defeats the purpose of even setting
the option in the first place.

From what I can tell from the upstart documentation it doesn’t try to
kill an application if it doesn’t shutdown from the first signal, though
I haven’t verified that this is the case.