Wait indef. for a controlled shutdown for RedHat distros
Thought it would be a lot nicer to actually wait for Kafka to shutdown
when `controlled.shutdown.enable` is set to `true`, so that one can use
the `restart` action.

Unfortunately I had to write a simplified version of `killproc` to
actually get this to work. I hate init scripts.