Set `config_dir` and `log4j` attrs. in attributes file
So, I decided to move them back since I realised that it’s nice to
manipulate the log4j configuration before actually including the
default recipe. To get around the issue with setting the log file for
an appender I decided to use `lazy` values (either through the use of
`lazy` or by creating a lambda/proc), which I think turned out to be
a quite nice solution, and another benefit of keeping the attributes
in the attributes file is that all defaults (or at least close to all)
are defined in the same place.