Check ports open by started process
Rather than checking a specific port we check that the started process
is listening to *some* port. The main reason for doing this is that with
the latest Kafka release (v0.9.0.0) one can have several "listeners",
probably for their new SSL support.

Fixed some minor mishaps with the SysV scripts.